remove chi2s; unit tests
