+++ src/simtools/sim_events/production_comparison.py
@@ -8,6 +8,7 @@ import numpy as np
 
 from simtools.sim_events.reader import EventDataReader
 from simtools.utils import names
+from simtools.utils.general import resolve_file_patterns
 
 
 @dataclass
@@ -42,6 +43,87 @@ class ProductionEventMetrics:
         if self.simulated_event_count <= 0:
             return 0.0
         return self.triggered_event_count / self.simulated_event_count
+
+
+def parse_production_arguments(production_arguments):
+    """Parse repeated production arguments into validated descriptors.
+
+    Parameters
+    ----------
+    production_arguments : list[list[str]]
+        Repeated ``--production`` arguments in the shape ``[label, patterns]``.
+
+    Returns
+    -------
+    list[ProductionDescriptor]
+        Validated and normalized production descriptors.
+
+    Raises
+    ------
+    ValueError
+        If configuration is malformed or does not contain any production.
+    """
+    parsed_productions = _normalize_production_arguments(production_arguments)
+    if not parsed_productions:
+        raise ValueError("At least one production is required.")
+
+    labels = [label for label, _ in parsed_productions]
+    if len(set(labels)) != len(labels):
+        raise ValueError("Production labels must be unique.")
+
+    descriptors = []
+    for label, pattern_list in parsed_productions:
+        patterns = [pattern.strip() for pattern in pattern_list.split(",") if pattern.strip()]
+        if len(patterns) == 0:
+            raise ValueError(f"Production '{label}' has no event_data_file pattern.")
+
+        resolved_files = [str(path) for path in resolve_file_patterns(patterns)]
+        if len(resolved_files) == 0:
+            raise ValueError(f"Production '{label}' does not resolve to any files.")
+        descriptors.append(ProductionDescriptor(label=label, event_data_files=resolved_files))
+
+    return descriptors
+
+
+def _normalize_production_arguments(production_arguments):
+    """Normalize raw production arguments into ``[(label, files), ...]``."""
+    if not production_arguments:
+        return []
+
+    normalized = []
+    if all(isinstance(item, str) for item in production_arguments):
+        return _pairwise_label_file_arguments(production_arguments)
+
+    for item in production_arguments:
+        normalized.extend(_normalize_single_production_argument(item))
+
+    return normalized
+
+
+def _pairwise_label_file_arguments(flat_arguments):
+    """Convert a flat list of strings into ``[(label, files), ...]`` pairs."""
+    if len(flat_arguments) % 2 != 0:
+        _raise_invalid_production_arguments()
+    return [
+        (flat_arguments[index], flat_arguments[index + 1])
+        for index in range(0, len(flat_arguments), 2)
+    ]
+
+
+def _normalize_single_production_argument(argument):
+    """Normalize one nested production argument into label/file pairs."""
+    if not isinstance(argument, list | tuple):
+        _raise_invalid_production_arguments()
+    if not all(isinstance(value, str) for value in argument):
+        _raise_invalid_production_arguments()
+    if len(argument) == 2:
+        return [(argument[0], argument[1])]
+    return _pairwise_label_file_arguments(list(argument))
+
+
+def _raise_invalid_production_arguments():
+    """Raise a standardized parser error for malformed production arguments."""
+    raise ValueError("Production arguments must be provided as label/file pairs.")
 
 
 def collect_production_metrics(production_descriptors, telescope_list=None):
